chore(suite-native): Refactor after rebase
