--- suite-native/graph/src/index.ts
@@ -1,2 +1,0 @@
-export * from './components/Graph';
-export * from './appGraphReducer';
