+++ suite-native/module-settings/src/index.ts
@@ -1,3 +1,3 @@
 export * from './navigation/routes';
 export * from './navigation/SettingsStackNavigator';
-export * from './appSettingsReducer';
+export * from './slice';
