+++ suite-native/state/src/store.ts
@@ -1,7 +1,7 @@
 import { configureStore, Store } from '@reduxjs/toolkit';
 
 import { appSettingsReducer } from '@suite-native/module-settings';
-import { appGraphReducer } from '@suite-native/graph';
+import { appGraphReducer } from '@suite-native/home-graph';
 
 export const store: Store = configureStore({
     reducer: {
